maint: Relicensed the manual from GFDL to GPLv3+
After adding content from the SIMP package, which is licensed under
GPLv3+, I had to consider whether this could be included in a manual
under GFDL. Also I found out that Debian does not consider GFDL as
free. Thus I investigated the differences between the licenses and
found out that the GPLv3 is more what I want for the package's manual.